Initial Commit: Done
At first, I thought this would be an easy assignment, and it was (compared to writing an interpreter from scratch in compiler theory).
What I didn't count on was running Clean and pulling my hair out trying to get the project to work again. I guess scenarios like that and publishing the code through GIT was the real point of this exercise. For what it's worth, I think this was a good surprise lesson.